+++ CHANGES.md
@@ -1,6 +1,13 @@
 # CHANGELOG
 
-## 0.6.1 (???)
+## 0.7.0 (???)
+
+### Improvements
+
+-   Calculate feature count by iterating over features if GDAL returns an
+    unknown count for a data layer (e.g., OSM driver); this may have signficant
+    performance impacts for some data sources that would otherwise return an
+    unknown count (count is used in `read_info`, `read`, `read_dataframe`).
 
 ### Bug fixes